changed Integer to Float for Results

+++ svm.logic.implementation/src/svm/logic/implementation/controller/ContestController.java
@@ -178,7 +178,7 @@ public class ContestController implements IContestController {
     }
 
     @Override
-    public void addResult(ITransferMatch match, Integer home, Integer away) throws RemoteException, NoSessionFoundException, DomainException, IllegalAccessException, InstantiationException {
+    public void addResult(ITransferMatch match, Float home, Float away) throws RemoteException, NoSessionFoundException, DomainException, IllegalAccessException, InstantiationException {
         IMatch m = ((IHasModel<IMatch>) match).getModel();
         m.addResult(home, away);
     }
